+++ src/main/java/com/github/morningzeng/toolset/utils/StringUtils.java
@@ -1,0 +1,13 @@
+package com.github.morningzeng.toolset.utils;
+
+/**
+ * @author Morning Zeng
+ * @since 2024-05-21
+ */
+public class StringUtils {
+
+    public static String maskSensitive(final String str) {
+        return str.replaceAll("(?<=.{3}).+(?=.{4})", "****");
+    }
+
+}